Moved validateRedirectUrl to src/lib/functions/urls.ts
