+++ src/frontend/src/lib/functions/urls.ts
@@ -13,3 +13,21 @@ export function strip_trailing_slash(input: string) {
 
 	return normalizedPath + (query ? '?' + query : '') + (hash ? '#' + hash : '');
 }
+
+export function validateRedirectUrl(url: string, origin: string): string {
+	try {
+		const parsed = new URL(url, origin);
+
+		if (parsed.origin !== origin) {
+			throw new Error('External redirects are not allowed.');
+		}
+
+		if (parsed.protocol !== 'http:' && parsed.protocol !== 'https:') {
+			throw new Error('Invalid protocol.');
+		}
+		return parsed.pathname + parsed.search + parsed.hash;
+	} catch (e) {
+		if (e instanceof Error) throw e;
+		throw new Error('Malformed redirect URL.');
+	}
+}
